requirements.txt
Django==5.2.4
djangorestframework==3.14.0
gunicorn==21.2.0
